complete payment exercise
Add ability to create payments for a loan using JSON API call.
Commented the protection from forgery to make this easier to test using
curl or something similar.

Expose outstanding balance for a loan in `LoanController#index` and
`LoanController#show`

Validate for creation of payments: payments cannot be created if they
would exceed the outstanding balance of a loan

Outstanding balance of a loan is calculated on the fly by summing all
of its payments' payment amounts. The balance of a loan is not stored in
the database.